chore(funcionario): update `TipoDocumentoId` fields to `Long` in `AnexoReqDTO` and `AgregadoDependenteReqDTO` including corresponding JSON definitions

+++ src/main/java/cv/inps/rh/funcionario/application/dto/AgregadoDependenteReqDTO.java
@@ -24,7 +24,7 @@ public class AgregadoDependenteReqDTO  {
   private Long id ;
   
   
-  private Integer tipoDocumentoId ;
+  private Long tipoDocumentoId ;
   
   
   private String numDocumento ;
